Working with examples and genai
The examples_prompt_buil.ipynb file is the file from the Nebari server which contains the functions I used to generate prompts designed to use few-shot prompting to generate examples. The examples_testing.txt goes through all of the tests that I went through. The results were very mixed, but the last five or so were functions that did not contain examples. I pulled the functions from core.py and _linalg.py from numpy. Not all of the examples generated were accurate and I am going to review the results more in depth for accuracy.