Adding the example_generator as well as output
The example_generator.ipynb is the generator I've been using to generate examples. The examples_testing_70.txt file contains the output from running the markdown document as written. I've tried running all of the code and have noticed that the output of running the code is often very close, however I've run into an issue with the svdvals function where I can seem to find how to run svdvals. I substitutied this with svd(a, compute_uv=False, hermitian=False) in order to evaluate the AI generated text.  Also the AI seemed to have imported the functions appropriately besides the svdvals function.