Created .gitignore file
This takes care of most autogenerated files from OS and Nebari.
It will capture .log files.
Feel free to add to this file if you create a new folder
that you don't want to be added to the server as it can be
generated with code.  Maybe we can adapt a generator type
thing as done on NumPy.