DOC: Generated Example Logs and Tracking List Logs
The tracking list logs are useful to store in the repo, as they make it
easy to grab information in other places. They can be generated from
list_callables.py, but I felt like leaving them in the repo was useful.

The logs from running example-generator.py take 2-10 minutes each to
generate, so storing them in the repo makes perfect sense.

I'm almost done running the example-generator on every function in numpy
except those in the base np space.