ENH: example_post_processing file began
This adds two new functions:
- search_and_replace_phrase for replacing old docstrings with updated.
- ppend_to_section for adding content to the end of section in a docstring.
These are two tools needed for processing the example logs generated by Llama3.
This is part of #77. I'll be adding a few more tools to this file before it's ready
to run on the entire Numpy codebase, which we can do one module at a time.