Mostly working replacer, with lots of new things to test.
I think it's time to start writing unit tests on the functions. I've been doing them
all along, but never named them or saved them as individual functions.
Right now, the code injects over 70% of the examples directly into the codebase
for the linalg module.  Errors are just printed directly in the code (ready for a
human to delete). More can be automated, and then it's time to run it on the entire codebase.